Make the online-phase process diagram match function definitions.
87a13127aa3cf0e0ea1d23a9640f06eb18d925fb changed the signature of
`CreateCredentialResponse`.

I also improved upon @kevinlewi's update.

I took the definitions of the OPRF functions and types from
https://cfrg.github.io/draft-irtf-cfrg-voprf/draft-irtf-cfrg-voprf.html .
I modified them some. Here is the original text for reference:
- SerializedElement and SerializedScalar are serialized representations of
  Element and Scalar types of length Ne and Ns, respectively
- SerializeElement(A): A member function of GG that maps a group element A to
  a unique byte array buf of fixed length Ne.
- DeserializeElement(buf): A member
  function of GG that maps a byte array buf to a group element A, or fails if
  the input is not a valid byte representation of an element.

I removed the references to `SerializeElement` and `DeserializeElement` since
they were not referenced anywhere in the draft.

There was also no mention of `SerializeScalar`. I added
"`kU` is serialized using `SerializeScalar`" to make its usage explicit.

I changed `CredentialRequest` and `CredentialResponse` to use
`SerializedElement` to match `RegistrationRequest` and `RegistrationResponse`.

In `RegistrationUpload` I moved `Envelope` to the end of the struct to be
consistent with the `Envelope`'s position in other structs:
```
struct {
    SerializedScalar kU;
    opaque pkU<1..2^16-1>;
    Envelope envU;
} CredentialFile;

struct {
    SerializedElement data;
    opaque pkS<1..2^16-1>;
    Envelope envU;
} CredentialResponse;
```
Also, `Envelope` is of variable length.