Expose export key after generating ke3
The client is meant to get an export key (in addition to the session key) from the registration flow but `generate_ke3` ignores the export key obtained by calling `core.recover_credentials` so it can't be retrieved. I save it to the instance so it can be retrieved in the same way as the session key.